authentication-service/requirements.txt
Flask
Flask-PyMongo
python-dotenv
Werkzeug
PyJWT
